backend/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework
djangorestframework-jsonapi
django-filter
pytest
pre-commit
black
